rename _types to types after uv 0.5.31
